Rearrange Cholesky impl/api headers
